pyproject.toml
[project]
name = "mnemonic"
version = "1.6.0"
description = "Persistent filesystem-based memory system for Claude Code"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "zircote"}
]
keywords = ["claude-code", "plugin", "memory", "ai", "mif"]
dependencies = [
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18.0",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/zircote/mnemonic"
Documentation = "https://github.com/zircote/mnemonic#readme"
Repository = "https://github.com/zircote/mnemonic.git"
Issues = "https://github.com/zircote/mnemonic/issues"

[tool.ruff]
target-version = "py38"
line-length = 120
exclude = ["docs/archive"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E402"]  # imports after sys.path.insert are expected in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
